+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
   - Don't show archived NOFOs on NOFO list page
   - Add warning banner for archived NOFOs to "view" and "edit" pages
   - Deleting NOFOs now just archives them
+  - Archived NOFOs can only be seen by Bloom users
 
 ### Changed